+++ src/fields/ColourSwatches.php
@@ -51,7 +51,7 @@ class ColourSwatches extends Field implements PreviewableFieldInterface
      */
     public static function displayName(): string
     {
-        return Craft::t('colour-swatches', 'Colour Swatches');
+        return Craft::t('colour-swatches', 'Color Swatches');
     }
 
     // Public Methods
@@ -83,11 +83,16 @@ class ColourSwatches extends Field implements PreviewableFieldInterface
      */
     public function normalizeValue($value, ElementInterface $element = null)
     {
+
         if ($value instanceof ColourSwatchesModel) {
             return $value;
         }
-
-        return new ColourSwatchesModel($value);
+        // quick array transform so that we can ensure and `required fields` fire an error
+        $valueData = (array)json_decode($value);
+        // if we have actual data return model
+        if(count($valueData) > 0) {
+             return new ColourSwatchesModel($value);
+        }
     }
 
     /**
@@ -95,11 +100,7 @@ class ColourSwatches extends Field implements PreviewableFieldInterface
      */
     public function serializeValue($value, ElementInterface $element = null)
     {
-        if ($value instanceof ColourSwatchesModel) {
-            return $value;
-        }
-
-        return new ColourSwatchesModel($value);
+        return $value;
     }
 
     /**
@@ -114,7 +115,7 @@ class ColourSwatches extends Field implements PreviewableFieldInterface
         } elseif ($this->palette) {
             $rows = Plugin::$plugin->settings->palettes[$this->palette];
         } else {
-            $rows = Plugin::$plugin->settings->colours;
+            $rows = Plugin::$plugin->settings->colors ? Plugin::$plugin->settings->colors : Plugin::$plugin->settings->colours;
         }
 
         $config = [
@@ -172,7 +173,7 @@ class ColourSwatches extends Field implements PreviewableFieldInterface
             [
                 'field'             => $this,
                 'config'            => $config,
-                'configOptions'     => Plugin::$plugin->settings->colours,
+                'configOptions'     => Plugin::$plugin->settings->colors ? Plugin::$plugin->settings->colors : Plugin::$plugin->settings->colours,
                 'paletteOptions'    => $paletteOptions,
                 'palettes'          => Plugin::$plugin->settings->palettes,
             ]
@@ -185,6 +186,7 @@ class ColourSwatches extends Field implements PreviewableFieldInterface
     public function getInputHtml($value, ElementInterface $element = null): string
     {
         // Register our asset bundle
+      //  print_r($value);
         Craft::$app->getView()->registerAssetBundle(ColourSwatchesFieldAsset::class);
 
         // Get our id and namespace
@@ -202,7 +204,7 @@ class ColourSwatches extends Field implements PreviewableFieldInterface
                 'field'        => $this,
                 'id'           => $id,
                 'namespacedId' => $namespacedId,
-                'configOptions'=> Plugin::$plugin->settings->colours,
+                'configOptions'=> Plugin::$plugin->settings->colors ? Plugin::$plugin->settings->colors : Plugin::$plugin->settings->colours,
                 'palettes'     => Plugin::$plugin->settings->palettes,
             ]
         );
@@ -212,19 +214,32 @@ class ColourSwatches extends Field implements PreviewableFieldInterface
      * {@inheritdoc}
      */
     public function getTableAttributeHtml($value, ElementInterface $element): string
-    {
-        /** @var ColorData|null $value */
-        $fieldValue;
-        if (!$value) {
-            return '<div class="color small static"><div class="color-preview"></div></div>';
+        {
+        // our preview no data value
+        $color = null;
+        $style="background-color: transparent";
+        // if we have data
+        if(!empty($value)){
+            $fieldValue = get_object_vars($value);
+            $gradients = array();
+            // if we have a custom color config
+            if (is_array($value->color)) {
+                foreach ($value->color as $color){
+                    $gradients[] = $color->color;
+                }
+                // set a fallback if we only have one colour
+                $style="background-color:$gradients[0]";
+                // else build the gradient
+                  if(count($gradients) > 1){
+                      $gradients = implode(",", $gradients);
+                      $style="background: linear-gradient(to bottom right, $gradients);";
+                  }
+              // if we're using the CP values
+              } else {
+                   $color = $value->color;
+                   $style="background-color:$color";
+              }
         }
-        if (is_array($value->color)) {
-            $fieldValue = $value->color[0]->color;
-        } else {
-            $fieldValue = $value->color;
-        }
-        return "<div class='color small static'><div class='color-preview' style='background-color: {$fieldValue};'></div></div>";
-        // . "<div class='colorhex code'>{$value->label}</div>";
-
+        return '<div class="color small static"><div class="color-preview" style="'.$style.'"></div></div>';
     }
 }