+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.4.0 - 2020-9-11 Bugfix
+Migrate namespace from rias to percipioglobal after upgrade. Also converted config php to 'colors' instead of 'colours'. Still supports 'colours' for older versions.
+Fixes issues (#36, #50)
 
 ## 1.3.3 - 2020-8-19 Bugfix
 Fixed bugs introduced in last release by not checking the default configuration settings - Apologies. New features coming in next release!
@@ -11,12 +14,12 @@ Fixes issues (#44, #45, #46, #47)
 
 ## 1.3.2 - 2020-8-13 Bugfix
 ElementInterface updated to allow colour preview for custom arrays. Fixes issue #42
-fieldValue conditions for standard colour strings udpated to reflect null update in version 1.3.1. Fixes issue #43
+fieldValue conditions for standard colour strings updated to reflect null update in version 1.3.1. Fixes issue #43
 
 
 ## 1.3.1 - 2020-8-03 Bugfix
 Updated save function to ensure custom array fields remain selected on resave / edit
-Allowed swtch between config or palette option in field settings
+Allowed switch between config or palette option in field settings
 
 ## 1.3.0.1 - 2020-7-31 HOTFIX
 temp fixed broken settings page, expect more robust repair next week :)